campaigns and benefits sync with api, prices converted to decimal and floats

+++ database/migrations/2021_07_05_173121_create_campaigns_table.php
@@ -21,7 +21,7 @@ class CreateCampaignsTable extends Migration
             $table->integer('make_id')->unsigned()->nullable();
             $table->integer('model_id')->unsigned()->nullable();
             $table->string('type')->nullable();
-            $table->integer('amount')->nullable();
+            $table->decimal('amount',13,2)->nullable();
             $table->dateTime('beginning')->nullable();
             $table->dateTime('end')->nullable();
             $table->timestamps();
